% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-annotate.R
\name{crtl_genes}
\alias{crtl_genes}
\title{Download and set up control genes}
\usage{
crtl_genes(overwrite = TRUE)
}
\arguments{
\item{overwrite}{Should the file be redownloaded and reset up?}
}
\description{
This function sets up the control genes file from DepMap is
called by the `gimap_annotate()`
}
