% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_example_data}
\alias{get_example_data}
\title{Returns example data for package}
\usage{
get_example_data(which_data)
}
\arguments{
\item{which_data}{options are "count" or "meta"; specifies which example dataset should be returned}
}
\value{
the respective example data either as a data frame or a specialized
gimap_dataset depending on what was requested.
}
\description{
This function loads and returns example data for the package.
Which dataset is returned must be specified. Data will be downloaded from Figshare
the first time it is used.
}
\examples{
\donttest{

counts_timepoint <- get_example_data("count")
counts_treatment <- get_example_data("count_treatment")
gimap_timepoint_dataset <- get_example_data("gimap")
gimap_treatment_dataset <- get_example_data("gimap_treatment")
metadata <- get_example_data("meta")
annotation <- get_example_data("annotation")
}
}
