% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-calculate_gi.R
\name{gimap_rep_stats}
\alias{gimap_rep_stats}
\title{Do tests for each replicate --an internal function used by calc_gi() function}
\usage{
gimap_rep_stats(replicate, gi_calc_double, gi_calc_single)
}
\arguments{
\item{replicate}{a name of a replicate to filter out from gi_calc_adj}

\item{gi_calc_double}{a data.frame with adjusted double gi scores}

\item{gi_calc_single}{a data.frame with adjusted single gi scores}
}
\description{
Create results table that has t test p values
}
