% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-gi.R
\name{plot_targets_bar}
\alias{plot_targets_bar}
\title{Plots for genetic interactions for specific targets}
\usage{
plot_targets_bar(gimap_dataset, target1, target2, reps_to_drop = "")
}
\arguments{
\item{gimap_dataset}{A special dataset structure that is setup using has had gi scores calculated.}

\item{target1}{Name of the first target to be plotted e.g.}

\item{target2}{Name of the second target to be plotted e.g.}
}
\description{
This plot is for when you'd like to examine a target pair specifically.
}
\examples{
\dontrun{

gimap_dataset <- get_example_data("gimap") \%>\%
  gimap_filter() \%>\%
  gimap_annotate(cell_line = "HELA") \%>\%
  gimap_normalize(
    timepoints = "day"
  ) \%>\%
  calc_gi()

# To plot results, pick out two targets from the gi_score table
head(dplyr::arrange(gimap_dataset$gi_score, fdr))

# "NDEL1_NDE1" is top result so let's plot that
plot_targets_bar(gimap_dataset, target1 = "NDEL1", target2 = "NDE1")
}
}
