% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-gi.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plots for Genetic interactions}
\usage{
plot_volcano(
  gimap_dataset,
  facet_rep = TRUE,
  reps_to_drop = c("Day05_RepA_early")
)
}
\arguments{
\item{gimap_dataset}{A special dataset structure that is setup using has had gi scores calculated.}

\item{facet_rep}{Should the replicates be wrapped with facet_wrap()?}

\item{#'}{@param reps_to_drop Names of replicates that should be not plotted (Optional)}
}
\description{
This plot is meant to be functionally equivalent to Fig S5K (for HeLa, equivalent of Fig 3a for PC9).
Scatter plot of target-level observed versus expected CRISPR scores in the HeLa screen.
The solid line is the linear regression line for the negative control (single KO) pgRNAs,
while dashed lines indicate ± 2 residuals.
}
\examples{
\dontrun{

gimap_dataset <- get_example_data("gimap") \%>\%
  gimap_filter() \%>\%
  gimap_annotate(cell_line = "HELA") \%>\%
  gimap_normalize(
    timepoints = "day"
  ) \%>\%
  calc_gi()

# To plot results
plot_exp_v_obs_scatter(gimap_dataset, reps_to_drop = "Day05_RepA_early")
plot_rank_scatter(gimap_dataset, reps_to_drop = "Day05_RepA_early")
plot_volcano(gimap_dataset, reps_to_drop = "Day05_RepA_early")
}
}
