% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-annotate.R
\name{refresh_annotation}
\alias{refresh_annotation}
\title{Refresh the annotation files by redownloading them}
\usage{
refresh_annotation()
}
\value{
options for tpm_file, cn_file, and ctrl_genes_file are set to NULL.
}
\description{
This function will set annotation file options to NULL so files
will be re-downloaded
}
\examples{

refresh_annotation()

}
