% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_tbl}
\alias{save_tbl}
\title{Save tbl}
\usage{
save_tbl(tbl, out_dir = NULL, params = NULL)
}
\arguments{
\item{tbl}{a table you want to save}

\item{out_dir}{the file path you want to save it to}

\item{params}{any additional parameters needed}
}
\description{
This function saves table
}
