% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-annotate.R
\name{supported_cell_lines}
\alias{supported_cell_lines}
\title{List the supported cell lines}
\usage{
supported_cell_lines()
}
\description{
This function downloads the metadata for DepMap and lists which cell
lines are supported.
}
\examples{
\dontrun{

cell_lines <- supported_cell_lines()
}
}
