% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-annotate.R
\name{tpm_setup}
\alias{tpm_setup}
\title{Download and set up DepMap TPM data}
\usage{
tpm_setup(overwrite = TRUE)
}
\arguments{
\item{overwrite}{should the files be re downloaded}
}
\description{
This function sets up the tpm data from DepMap is
called by the `gimap_annotate()` function
}
